# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "DoomsdAPI",
    "description" : "This is service is designed to return current information and historical data related to the Doomsday Clock maintained by the Bulletin of Atomic Scientists. There is a lightweight front end application designed to emmulate the motif of the clock as described here - https://www.theatlantic.com/entertainment/archive/2015/11/doomsday-clock-michael-bierut-design/412936/",
    "contact" : {
      "email" : "nate01776@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "tags" : [ {
    "name" : "Consumer",
    "description" : "Operations available to API consumers"
  }, {
    "name" : "Provider",
    "description" : "Operations specific to API provider"
  } ],
  "paths" : {
    "/time" : {
      "get" : {
        "tags" : [ "Consumer", "Provider" ],
        "summary" : "Current status of clock.",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "clockStatus",
        "parameters" : [ {
          "name" : "searchString",
          "in" : "query",
          "description" : "pass an optional search string for looking up inventory",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Current status of clock.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StatusModel"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      }
    },
    "/history" : {
      "get" : {
        "tags" : [ "Consumer", "Provider" ],
        "summary" : "History of clocks status",
        "description" : "Returns a history of the clock's time to midnight with optional descriptions",
        "operationId" : "clockHistory",
        "parameters" : [ {
          "name" : "startDate",
          "in" : "query",
          "description" : "Hyphen seperated start date from which status objects are returned.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "example" : "05-05-1947"
          }
        }, {
          "name" : "endDate",
          "in" : "query",
          "description" : "Hyphen seperated end date of history window, if not included default will be the current date",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "example" : "01-28-2020"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Array of status objects describing the various changes the clock has gone through, scoped based on submitted parameters.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StatusModel"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      }
    },
    "/data/{type}" : {
      "get" : {
        "tags" : [ "Consumer", "Provider" ],
        "summary" : "History of clocks status",
        "description" : "Returns a",
        "operationId" : "dataQuery",
        "parameters" : [ {
          "name" : "type",
          "in" : "path",
          "description" : "Defines the type of influence data, provider directly from the Bulletin of Atomic Scientist's sources.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "enum"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Array of status objects describing the various changes the clock has gone through, scoped based on submitted parameters.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StatusModel"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Server Error"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "StatusModel" : {
        "required" : [ "date", "id", "longMessage", "shortMessage", "timeToMidnight" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "Internal reference for a clock status.",
            "example" : 3
          },
          "date" : {
            "type" : "string",
            "description" : "Date a TTM change was released",
            "format" : "date-time"
          },
          "timeToMidnight" : {
            "type" : "string",
            "description" : "Time to midnight",
            "example" : "100 seconds"
          },
          "shortMessage" : {
            "type" : "string",
            "description" : "Short message including the clocks TTM,",
            "example" : "It is 100 seconds to midnight."
          },
          "ref" : {
            "type" : "string",
            "description" : "URL reference to the Bulletin of Atomic Scientist's post regarding this status.",
            "example" : "https://thebulletin.org/doomsday-clock/current-time/"
          },
          "longMessage" : {
            "type" : "string",
            "description" : "Descriptive message about the TTM status."
          }
        },
        "example" : {
          "date" : "2000-01-23T04:56:07.000+00:00",
          "ref" : "https://thebulletin.org/doomsday-clock/current-time/",
          "timeToMidnight" : "100 seconds",
          "shortMessage" : "It is 100 seconds to midnight.",
          "id" : 3,
          "longMessage" : "longMessage"
        }
      }
    }
  }
}